requirements.txt
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.33.0
Pillow>=10.0.0
numpy>=1.25.0
